oval/roundrect cleanup
