+++ src/quickdraw/qStdOval.cpp
@@ -10,8 +10,146 @@
 #include <MemoryMgr.h>
 #include <quickdraw/cquick.h>
 #include <quickdraw/picture.h>
+#include <util/handle_vector.h>
 
 using namespace Executor;
+
+
+using RgnVector = handle_vector<int16_t, RgnHandle, 10>;
+
+std::vector<int16_t> bresenham(int64_t width, int64_t height)
+{
+    std::vector<int16_t> steps;
+    steps.reserve(height);
+
+    int64_t a = int64_t(height) * height;
+    int64_t b = int64_t(width) * width;
+
+    bool oddWidth = width % 2;
+    bool oddHeight = height % 2;
+
+    int64_t d = (oddWidth ? 4 : 1) * height * height + (1 - 2 * height) * width * width;
+
+    int x = 0, y = height/2;
+
+    int64_t deltaD;
+    do
+    {
+        while(d < 0)
+        {
+            d += 4 * a * (2 * x + 1 + (oddWidth ? 2 : 1));
+            ++x;
+        }
+        
+        steps.push_back(x);
+        deltaD = 4 * a * (2 * x + 1 + (oddWidth ? 2 : 1)) - 4 * b * (2 * y - (oddHeight ? 1 : 2));
+        d += deltaD;
+        --y;
+        ++x;
+    } while(deltaD <= 0);
+    
+    return steps;
+}
+
+RgnHandle Executor::ROMlib_circrgn(const Rect& rect)
+{
+    RgnVector rgn;
+
+    int64_t width = int(rect.right) - rect.left;
+    int64_t height = int(rect.bottom) - rect.top;
+
+    if(width >= 3 && height >= 3)
+    {
+        auto xsteps = bresenham(width, height);
+        auto ysteps = bresenham(height, width);
+
+#if 0
+        std::cout << "CIRCLE " << width << "x" << height << std::endl;
+        for(auto x : xsteps)
+            std::cout << " " << x;
+        std::cout << "\n";
+        for(auto y : ysteps)
+            std::cout << " " << y;
+        std::cout << "\n";
+#endif
+
+        int16_t centl = rect.left + width / 2;
+        int16_t centr = rect.right - width / 2;
+        int16_t centt = rect.top + height / 2;
+        int16_t centb = rect.bottom - height / 2;
+
+        int16_t ox = -1;
+
+        auto pointUpper = [&ox, &rgn, centl, centr](int16_t x, int16_t y) {
+            if(x == ox)
+                return;
+            rgn.push_back(y);
+            rgn.push_back(centl - x);
+            if(ox >= 0)
+            {
+                rgn.push_back(centl - ox);
+                rgn.push_back(centr + ox);
+            }
+            rgn.push_back(centr + x);
+            rgn.push_back(RGN_STOP);
+            ox = x;
+        };
+        auto pointLower = [&ox, &rgn, centl, centr](int16_t x, int16_t y) {
+            if(x == ox)
+                return;
+            rgn.push_back(y);
+            rgn.push_back(centl - ox);
+            rgn.push_back(centl - x);
+            rgn.push_back(centr + x);
+            rgn.push_back(centr + ox);
+            rgn.push_back(RGN_STOP);
+            ox = x;
+        };
+
+        {
+            int16_t y = rect.top;
+            
+            for(int16_t x : xsteps)
+                pointUpper(x, y++);
+        }
+            
+        for(int i = ysteps.size() - 1; i >= 0; i--)
+        {
+            int16_t x = width / 2 - i;
+            int16_t y = centt - ysteps[i];
+
+            pointUpper(x, y);
+        }
+
+        for(int i = 0; i < ysteps.size()-1; i++)
+        {
+            int16_t x = width / 2 - i - 1;
+            int16_t y = centb + ysteps[i];
+
+            pointLower(x, y);
+        }
+
+        for(int i = xsteps.size() - 1; i >= 0; i--)
+        {
+            int16_t y = rect.bottom - i - 1;
+            int16_t x = xsteps[i];
+
+            pointLower(x, y);
+        }
+
+        rgn.push_back(rect.bottom);
+        rgn.push_back(centl - ox);
+        rgn.push_back(centr + ox);
+        rgn.push_back(RGN_STOP);
+        rgn.push_back(RGN_STOP);
+    }
+
+    size_t sz = sizeof(int16_t) * rgn.size() + 10;
+    RgnHandle rgnH = rgn.release();
+    (*rgnH)->rgnBBox = rect;
+    (*rgnH)->rgnSize = sz;
+    return rgnH;
+}
 
 void Executor::C_StdOval(GrafVerb v, const Rect *rp)
 {
@@ -33,7 +171,7 @@ void Executor::C_StdOval(GrafVerb v, const Rect *rp)
             StdRect(v, rp);
         else
         {
-            rh = ROMlib_circrgn(rp);
+            rh = ROMlib_circrgn(*rp);
             switch(v)
             {
                 case frame:
@@ -49,7 +187,7 @@ void Executor::C_StdOval(GrafVerb v, const Rect *rp)
                         r.right = rp->right - PORT_PEN_SIZE(qdGlobals().thePort).h;
                         if(r.top < r.bottom && r.left < r.right)
                         {
-                            rh2 = ROMlib_circrgn(&r);
+                            rh2 = ROMlib_circrgn(r);
                             XorRgn(rh, rh2, rh);
                             DisposeRgn(rh2);
                         }
